fix(review): address H2-H4, all M and L1/L4/L5/L6 findings
- H2: subscribeTopic idempotent via contains check (WebSocketConnectionManager)
- H3: monotonic System.nanoTime() for cooldown/heartbeat/idle intervals
  (BackpressureManager, ClusterMeshManager, WebSocketConnectionManager)
- H4: backpressure release-check self-rescheduling retry loop, dedup flag,
  registration-time writability check, null eventLoop guard
- M1: deprecate WebSocketConnectionManager, production path unified in PacketRouter
- M2: log pending count before clear in ResponseMapping.shutdown
- M3: request id int->AtomicLong to avoid 2^31 wrap collision
- M4: TcpConnectionPool keeps LongAdder counters when array empties (no race re-create)
- M5: broadcast/broadcastToTopic per-channel try-catch + slice release on write failure
- M6: per-channel topic reverse index for O(k) disconnect cleanup
- M7: configurable idle timeout/write water marks/sndbuf/rcvbuf in WebSocketConfig
- M8: base-node connect retry loop instead of connect-once after blind sleep
- L1: remove duplicated JavaDoc blocks (TcpClient)
- L4: drop redundant topic null check (PacketRouter.broadcastToTopic)
- L5: drop unused imports (ResponseMapping)
- L6: document scheduledEventLoop as non-sync retry channel hint (BackpressureManager)
- H1 deferred: README/README_cn note intranet-only deployment, auth not enforced
- Test: ResponseMappingTest adapted to long request ids; full suites match pre-existing baseline failures

+++ jwsch-test/src/test/java/cn/itcraft/jwsch/srv/router/ResponseMappingTest.java
@@ -24,8 +24,8 @@ public class ResponseMappingTest {
     
     @Test
     public void testGenerateRequestId() {
-        int id1 = responseMapping.generateRequestId();
-        int id2 = responseMapping.generateRequestId();
+        long id1 = responseMapping.generateRequestId();
+        long id2 = responseMapping.generateRequestId();
         
         assertNotEquals(id1, id2);
         assertTrue(id1 > 0);
@@ -33,7 +33,7 @@ public class ResponseMappingTest {
     
     @Test
     public void testCreateFuture() {
-        int requestId = responseMapping.generateRequestId();
+        long requestId = responseMapping.generateRequestId();
         CompletableFuture<Packet> future = responseMapping.createFuture(requestId);
         
         assertNotNull(future);
@@ -43,7 +43,7 @@ public class ResponseMappingTest {
     
     @Test
     public void testCompleteResponse() {
-        int requestId = responseMapping.generateRequestId();
+        long requestId = responseMapping.generateRequestId();
         CompletableFuture<Packet> future = responseMapping.createFuture(requestId);
         
         Packet response = createTestPacket();
@@ -63,7 +63,7 @@ public class ResponseMappingTest {
     
     @Test
     public void testTimeout() throws Exception {
-        int requestId = responseMapping.generateRequestId();
+        long requestId = responseMapping.generateRequestId();
         CompletableFuture<Packet> future = responseMapping.createFuture(requestId);
         
         try {
@@ -79,7 +79,7 @@ public class ResponseMappingTest {
     
     @Test
     public void testRemoveFuture() {
-        int requestId = responseMapping.generateRequestId();
+        long requestId = responseMapping.generateRequestId();
         responseMapping.createFuture(requestId);
         
         responseMapping.removeFuture(requestId);
